+++ lib/px/pxconfig.lua
@@ -10,24 +10,28 @@ _M.px_enabled = true
 
 -- ##  Configuration Block ##
 
+
 _M.px_appId = 'PX_APP_ID'
 _M.cookie_secret = 'COOKIE_SECRET'
 _M.auth_token = 'PX_AUTH_TOKEN'
 _M.blocking_score = 60
-_M.cookie_encrypted = true
-_M.enable_server_calls = true
-_M.send_page_requested_activity = true
+_M.px_debug = false
 _M.block_enabled = true
 _M.captcha_enabled = true
-_M.px_debug = true
+_M.sensitive_headers = {'cookie', 'cookies'}
+_M.ip_headers = {}
+_M.s2s_timeout = 1000
+_M.client_timeout = 2000
+_M.cookie_encrypted = true
+_M.send_page_requested_activity = true
+_M.score_header_name = 'X-PX-SCORE'
 _M.sensitive_routes_prefix = {}
 _M.sensitive_routes_suffix = {}
 _M.captcha_provider = "reCaptcha"
 _M.additional_activity_handler = nil
+_M.enable_server_calls = true
 
-_M.s2s_timeout = 1000
 _M.px_maxbuflen = 10
-_M.score_header_name = 'X-PX-SCORE'
 _M.px_port = 443
 _M.ssl_enabled = true
 _M.custom_block_url = nil
@@ -35,7 +39,7 @@ _M.enabled_routes = {}
 _M.custom_logo = nil
 _M.css_ref = nil
 _M.js_ref = nil
--- -- ## END - Configuration block ##
+-- ## END - Configuration block ##
 
 -- ## Filter Configuration ##