+++ lib/px/pxnginx.lua
@@ -25,35 +25,36 @@
 local M = {}
 
 function M.application(file_name)
-    local _M = {}
-    -- Support for multiple apps - each app file should be named "pxconfig-<appname>.lua"
     local config_file = ((file_name == nil or file_name == '') and "px.pxconfig" or "px.pxconfig-" .. file_name)
 
     local px_config = require(config_file)
+    local _M = {}
+    -- Support for multiple apps - each app file should be named "pxconfig-<appname>.lua"
+
     local px_filters = require("px.utils.pxfilters").load(config_file)
     local px_client = require("px.utils.pxclient").load(config_file)
-    local px_cookie_v1 = require("px.utils.pxcookie_v1").load(config_file)
-    local px_cookie_v3 = require("px.utils.pxcookie_v3").load(config_file)
+    local PXPayload = require('px.utils.pxpayload')
+    local px_payload = PXPayload:new{}
     local px_captcha = require("px.utils.pxcaptcha").load(config_file)
     local px_block = require("px.block.pxblock").load(config_file)
     local px_api = require("px.utils.pxapi").load(config_file)
     local px_logger = require("px.utils.pxlogger").load(config_file)
     local px_headers = require("px.utils.pxheaders").load(config_file)
     local px_constants = require("px.utils.pxconstants")
-    local px_common_utils = require("px.utils.pxcommonutils").load()
+    local px_common_utils = require("px.utils.pxcommonutils")
 
     local auth_token = px_config.auth_token
     local enable_server_calls = px_config.enable_server_calls
     local risk_api_path = px_constants.RISK_PATH
     local enabled_routes = px_config.enabled_routes
-    local remote_addr = ngx.var.remote_addr or ""
+    local remote_addr = px_headers.get_ip()
     local user_agent = ngx.var.http_user_agent or ""
     local string_sub = string.sub
     local string_len = string.len
     local pcall = pcall
 
     local function perform_s2s(result, details)
-	ngx.ctx.s2s_call_reason = result.message
+        ngx.ctx.s2s_call_reason = result.message
         local request_data = px_api.new_request_object(result.message)
         local start_risk_rtt = px_common_utils.get_time_in_milliseconds()
         local success, response = pcall(px_api.call_s2s, request_data, risk_api_path, auth_token)
@@ -126,7 +127,7 @@ function M.application(file_name)
     local _pxCaptcha = ngx.var.cookie__pxCaptcha
     local details = {};
 
-    if px_config.captcha_enabled and _pxCaptcha then
+    if _pxCaptcha then
         local success, result = pcall(px_captcha.process, _pxCaptcha)
 
         -- validating captcha value and if reset was successful then pass the request
@@ -138,65 +139,32 @@ function M.application(file_name)
         end
     end
 
-    -- process _px cookie v3 if present
-    local _px3 = ngx.var.cookie__px3
-    local _px = ngx.var.cookie__px
+    px_payload:load(config_file)
+    px_cookie = px_payload:get_payload()
+    px_cookie:load(config_file)
 
-    if _px3 then
-        ngx.ctx.px_orig_cookie = _px3
-        local success, result = pcall(px_cookie_v3.process, _px3)
-        -- cookie verification passed - checking result.
-        if success then
-            px_logger.debug("PX-CookieV3 Processed Succesfuly")
-            details["px_cookie"] = ngx.ctx.px_cookie;
-            details["px_cookie_hmac"] = ngx.ctx.px_cookie_hmac;
-            details["px_cookie_version"] = 'v3';
-            -- score crossed threshold
-            if result == false then
-                return px_block.block('cookie_high_score')
-                -- score did not cross the blocking threshold
-            else
-                ngx.ctx.pass_reason = 'cookie'
-                px_client.send_to_perimeterx("page_requested", details)
-                return true
-            end
-            -- cookie verification failed/cookie does not exist. performing s2s query
-        elseif enable_server_calls == true then
-            return perform_s2s(result, details)
+    local success, result = pcall(px_cookie.process, px_cookie)
+    -- cookie verification passed - checking result.
+    if success then
+        px_logger.debug("PX-Cookie Processed Successfully")
+        details["px_cookie"] = ngx.ctx.px_cookie;
+        details["px_cookie_hmac"] = ngx.ctx.px_cookie_hmac;
+        details["px_cookie_version"] = ngx.ctx.px_cookie_version;
+
+        -- score crossed threshold
+        if result == false then
+            return px_block.block('cookie_high_score')
         else
-            ngx.ctx.pass_reason = 'error'
+            ngx.ctx.pass_reason = 'cookie'
             px_client.send_to_perimeterx("page_requested", details)
             return true
         end
+    elseif enable_server_calls == true then
+        return perform_s2s(result, details)
     else
-        -- process _px cookie v1
-        ngx.ctx.px_orig_cookie = _px
-        local success, result = pcall(px_cookie_v1.process, _px)
-        local details = {};
-        -- cookie verification passed - checking result.
-        if success then
-            px_logger.debug("PX-CookieV1 Processed Succesfuly")
-            details["px_cookie"] = ngx.ctx.px_cookie;
-            details["px_cookie_hmac"] = ngx.ctx.px_cookie_hmac;
-            details["px_cookie_version"] = 'v1';
-            -- score crossed threshold
-            if result == false then
-                ngx.ctx.px_action = 'c'
-                return px_block.block('cookie_high_score')
-                -- score did not cross the blocking threshold
-            else
-                ngx.ctx.pass_reason = 'cookie';
-                px_client.send_to_perimeterx("page_requested", details)
-                return true
-            end
-            -- cookie verification failed/cookie does not exist. performing s2s query
-        elseif enable_server_calls == true then
-            return perform_s2s(result, details)
-        else
-            ngx.ctx.pass_reason = 'error'
-            px_client.send_to_perimeterx("page_requested", details)
-            return true
-        end
+        ngx.ctx.pass_reason = 'error'
+        px_client.send_to_perimeterx("page_requested", details)
+        return true
     end
 end