+++ lib/px/utils/pxconstants.lua
@@ -15,10 +15,10 @@ local function readonlytable(table)
 end
 
 local _M = readonlytable {
-    MODULE_VERSION = "NGINX Module v2.9.0",
+    MODULE_VERSION = "NGINX Module v2.10.0",
     RISK_PATH = "/api/v2/risk",
     CAPTCHA_PATH = "/api/v2/risk/captcha",
-    ACTIVITIES_PATH = "/api/v1/collector/s2s",
+    ACTIVITIES_PATH = "/api/v1/collector/s2s"
 }
 
 return _M